test: split lifecycle spool tests
